context.HttpContext

+++ Models/WeatherViewModel.cs
@@ -4,12 +4,12 @@ namespace WeatherAPI.Models
 {
     public class WeatherViewModel
     {
-        [JsonProperty("humidity")] public long Humidity { get; set; }
+        [JsonProperty("humidity")] public int Humidity { get; set; }
 
-        [JsonProperty("temperature")] public long Temperature { get; set; }
+        [JsonProperty("temperature")] public int Temperature { get; set; }
 
-        [JsonProperty("min_temperature")] public long MinTemperature { get; set; }
+        [JsonProperty("min_temperature")] public int MinTemperature { get; set; }
 
-        [JsonProperty("max_temperature")] public long MaxTemperature { get; set; }
+        [JsonProperty("max_temperature")] public int MaxTemperature { get; set; }
     }
 }
